requirements.txt
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.65.0
albumentations>=1.3.0
ultralytics>=8.0.0
opencv-python>=4.8.0
urllib3>=2.0.7
nicegui>=1.3.8
torch>=2.0.0
